+++ src/lib.rs
@@ -226,11 +226,21 @@ mod browse_result;
 #[cfg(feature = "tokio")]
 mod callback;
 mod data_value;
-mod logger;
+#[cfg(feature = "mbedtls")]
+mod ssl;
 mod traits;
 mod userdata;
 mod value;
 
+#[cfg(feature = "mbedtls")]
+pub use self::ssl::{create_certificate, Certificate};
+#[cfg(feature = "tokio")]
+pub use self::{
+    async_client::AsyncClient,
+    async_monitored_item::AsyncMonitoredItem,
+    async_subscription::AsyncSubscription,
+    callback::{CallbackOnce, CallbackStream},
+};
 pub use self::{
     browse_result::BrowseResult,
     client::{Client, ClientBuilder},
@@ -249,17 +259,8 @@ pub use self::{
 };
 pub(crate) use self::{
     data_type::{bitmask_ops, data_type, enum_variants},
-    logger::logger,
     service::{ServiceRequest, ServiceResponse},
     value::{ArrayValue, NonScalarValue},
-};
-
-#[cfg(feature = "tokio")]
-pub use self::{
-    async_client::AsyncClient,
-    async_monitored_item::AsyncMonitoredItem,
-    async_subscription::AsyncSubscription,
-    callback::{CallbackOnce, CallbackStream},
 };
 
 /// IANA-assigned OPC UA port number.