+++ src/ua.rs
@@ -3,11 +3,15 @@
 mod access_level;
 mod array;
 mod browse_result_mask;
+#[cfg(feature = "mbedtls")]
+mod certificate_format;
 mod client;
 mod client_config;
 mod continuation_point;
 mod data_types;
 mod event_id;
+mod key_value_map;
+mod logger;
 mod monitored_item_id;
 mod node_class_mask;
 mod secure_channel_state;
@@ -18,6 +22,8 @@ mod specified_attributes;
 mod subscription_id;
 mod user_identity_token;
 
+#[cfg(feature = "mbedtls")]
+pub use self::certificate_format::CertificateFormat;
 pub use self::{
     access_level::AccessLevel,
     array::Array,
@@ -26,6 +32,7 @@ pub use self::{
     continuation_point::ContinuationPoint,
     data_types::*,
     event_id::EventId,
+    key_value_map::KeyValueMap,
     monitored_item_id::MonitoredItemId,
     node_class_mask::NodeClassMask,
     secure_channel_state::SecureChannelState,
@@ -35,4 +42,4 @@ pub use self::{
     subscription_id::SubscriptionId,
     user_identity_token::UserIdentityToken,
 };
-pub(crate) use self::{client_config::ClientConfig, server_config::ServerConfig};
+pub(crate) use self::{client_config::ClientConfig, logger::Logger, server_config::ServerConfig};